@@ -14,30 +14,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.4",
-    "@angular/cdk": "^2.0.0-beta.12",
-    "@angular/common": "^4.4.4",
-    "@angular/compiler": "^4.4.4",
-    "@angular/core": "^4.4.4",
-    "@angular/flex-layout": "^2.0.0-beta.9",
-    "@angular/forms": "^4.4.4",
-    "@angular/http": "^4.4.4",
-    "@angular/material": "^2.0.0-beta.12",
-    "@angular/platform-browser": "^4.4.4",
-    "@angular/platform-browser-dynamic": "^4.4.4",
-    "@angular/router": "^4.4.4",
+    "@angular/animations": "^5.0.1",
+    "@angular/cdk": "^5.0.0-rc0",
+    "@angular/common": "^5.0.1",
+    "@angular/compiler": "^5.0.1",
+    "@angular/core": "^5.0.1",
+    "@angular/flex-layout": "^2.0.0-beta.10-4905443",
+    "@angular/forms": "^5.0.1",
+    "@angular/http": "^5.0.1",
+    "@angular/material": "^5.0.0-rc0",
+    "@angular/platform-browser": "^5.0.1",
+    "@angular/platform-browser-dynamic": "^5.0.1",
+    "@angular/router": "^5.0.1",
     "core-js": "^2.5.1",
     "hammerjs": "^2.0.8",
-    "rxjs": "^5.4.3",
+    "rxjs": "^5.5.2",
     "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "1.4.4",
-    "@angular/compiler-cli": "^4.4.4",
+    "@angular/cli": "1.5.0",
+    "@angular/compiler-cli": "^5.0.1",
     "@types/hammerjs": "^2.0.35",
-    "@types/jasmine": "2.6.0",
-    "@types/node": "~8.0.31",
-    "codelyzer": "~3.2.0",
+    "@types/jasmine": "2.6.3",
+    "@types/node": "~8.0.51",
+    "codelyzer": "~4.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~1.7.1",
@@ -47,10 +47,10 @@
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-phantomjs-launcher": "^1.0.4",
-    "phantomjs-prebuilt": "^2.1.15",
-    "protractor": "~5.1.2",
+    "phantomjs-prebuilt": "^2.1.16",
+    "protractor": "~5.2.0",
     "ts-node": "~3.3.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.5.3"
+    "tslint": "~5.8.0",
+    "typescript": "~2.6.1"
   }
 }
